+++ jl_sys/src/inlined.rs
@@ -1,33 +1,39 @@
 // If LTO is not enabled accessing arrays is very slow, so we're going to optimize
 // the common case a little.
 
-use std::ptr::NonNull;
+use std::{
+    marker::{PhantomData, PhantomPinned},
+    ptr::NonNull,
+};
 
 use crate::jl_array_t;
 
-#[cfg(any(feature = "julia-1-10"))]
+#[cfg(julia_1_10)]
 #[inline]
 pub const unsafe fn jlrs_array_data_fast(a: *mut jl_array_t) -> *mut std::ffi::c_void {
     #[repr(C)]
     struct RawArray {
         ptr: *mut std::ffi::c_void,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     NonNull::new_unchecked(a as *mut RawArray).as_ref().ptr
 }
 
-#[cfg(not(any(feature = "julia-1-10")))]
+#[cfg(not(julia_1_10))]
 #[inline]
 pub const unsafe fn jlrs_array_data_fast(a: *mut jl_array_t) -> *mut std::ffi::c_void {
     #[repr(C)]
     struct GenericMemoryRef {
         ptr_or_offset: *mut std::ffi::c_void,
         mem: *mut std::ffi::c_void,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     #[repr(C)]
     struct RawArray {
         ref_inner: GenericMemoryRef,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     NonNull::new_unchecked(a as *mut RawArray)
@@ -36,7 +42,7 @@ pub const unsafe fn jlrs_array_data_fast(a: *mut jl_array_t) -> *mut std::ffi::c
         .ptr_or_offset
 }
 
-#[cfg(any(feature = "julia-1-10"))]
+#[cfg(julia_1_10)]
 #[inline]
 pub const unsafe fn jlrs_array_dims_ptr(a: *mut jl_array_t) -> *mut usize {
     #[repr(C)]
@@ -47,42 +53,47 @@ pub const unsafe fn jlrs_array_dims_ptr(a: *mut jl_array_t) -> *mut usize {
         elsize: u16,
         offset: u32,
         nrows: usize,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     const OFFSET: usize = std::mem::offset_of!(RawArray, nrows);
     (a as *mut u8).add(OFFSET) as *mut usize
 }
 
-#[cfg(not(any(feature = "julia-1-10")))]
+#[cfg(not(julia_1_10))]
 #[inline]
 pub const unsafe fn jlrs_array_dims_ptr(a: *mut jl_array_t) -> *mut usize {
     #[repr(C)]
     struct GenericMemoryRef {
         ptr_or_offset: *mut std::ffi::c_void,
         mem: *mut std::ffi::c_void,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     #[repr(C)]
     struct RawArray {
         ref_inner: GenericMemoryRef,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     const OFFSET: usize = std::mem::size_of::<RawArray>();
     (a as *mut u8).add(OFFSET) as *mut usize
 }
 
-#[cfg(not(any(feature = "julia-1-10")))]
+#[cfg(not(julia_1_10))]
 #[inline]
 pub const unsafe fn jlrs_array_mem(a: *mut jl_array_t) -> *mut crate::types::jl_value_t {
     #[repr(C)]
     struct GenericMemoryRef {
         ptr_or_offset: *mut std::ffi::c_void,
         mem: *mut std::ffi::c_void,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     #[repr(C)]
     struct RawArray {
         ref_inner: GenericMemoryRef,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     NonNull::new_unchecked(a as *mut RawArray)
@@ -98,6 +109,7 @@ pub const unsafe fn jlrs_array_ndims_fast(a: *mut jl_array_t) -> usize {
         name: *mut std::ffi::c_void,
         super_ty: *mut std::ffi::c_void,
         parameters: *mut std::ffi::c_void,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     #[repr(C)]
@@ -106,11 +118,13 @@ pub const unsafe fn jlrs_array_ndims_fast(a: *mut jl_array_t) -> usize {
         next: *mut TaggedValue,
         ty: *mut RawDataType,
         bits: usize,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     #[repr(C)]
     struct TaggedValue {
         header: Header,
+        _marker: PhantomData<(*mut u8, PhantomPinned)>,
     }
 
     let a = a as *mut u8;