+++ jl_sys/src/types.rs
@@ -3,6 +3,7 @@
 use std::{
     cell::Cell,
     ffi::{c_int, c_void},
+    marker::{PhantomData, PhantomPinned},
     ptr::null_mut,
 };
 
@@ -18,24 +19,28 @@ pub enum jlrs_catch_tag_t {
 #[derive(Copy, Clone, Debug)]
 pub struct jl_array_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_datatype_layout_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_datatype_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_expr_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(u32)]
@@ -53,6 +58,7 @@ pub type GcCollection = jl_gc_collection_t;
 pub struct jl_gcframe_t {
     pub(crate) n_roots: usize,
     pub(crate) prev: Cell<*mut c_void>,
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 impl jl_gcframe_t {
@@ -61,6 +67,7 @@ impl jl_gcframe_t {
         jl_gcframe_t {
             n_roots: N << 2,
             prev: Cell::new(null_mut()),
+            _marker: PhantomData,
         }
     }
 
@@ -69,15 +76,17 @@ impl jl_gcframe_t {
         jl_gcframe_t {
             n_roots: (m + n) << 2,
             prev: Cell::new(null_mut()),
+            _marker: PhantomData,
         }
     }
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
-#[cfg(not(any(feature = "julia-1-10",)))]
+#[cfg(not(julia_1_10))]
 pub struct jl_genericmemory_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 pub type jl_markfunc_t =
@@ -91,6 +100,7 @@ pub type jl_gc_cb_root_scanner_t = unsafe extern "C" fn(c_int);
 #[derive(Copy, Clone, Debug)]
 pub struct jl_module_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
@@ -112,58 +122,68 @@ pub type jlrs_unsized_scope_trampoline_t = unsafe extern "C-unwind" fn(
 #[derive(Copy, Clone, Debug)]
 pub struct JL_STREAM {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_svec_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_sym_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_task_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_tls_states_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_tvar_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_typename_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_unionall_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_uniontype_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug)]
 pub struct jl_value_t {
     _unused: [u8; 0],
+    _marker: PhantomData<(*mut u8, PhantomPinned)>,
 }