+++ jlrs/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    jlrs_cfg::configure(10, 13);
+}